fix(DataType): Fix Data Type & Implement TestUtils
- fix number data type: Long.
- implement testUtils each domains.

+++ src/main/java/org/team6/coffeebeanery/order/dto/OrderDTO.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Setter
 @Builder
 public class OrderDTO {
-    private Integer orderId; //주문 ID
+    private Long orderId; //주문 ID
     
     private String customerEmail; //고객 이메일
     
@@ -23,7 +23,7 @@ public class OrderDTO {
     
     private LocalDateTime orderCreatedAt; //주문 날짜 및 시간
     
-    private Double totalPrice; //총 주문 금액
+    private Long totalPrice; //총 주문 금액
     
     private OrderStatus orderStatus; //주문 상태
